Remove animateProperty and move animateCubicBezier
