+++ src/BasicBehaveEngine/BasicBehaveEngine.ts
@@ -107,7 +107,7 @@ import { InverseHyperbolicTangent } from "./nodes/math/hyperbolic/InverseHyperbo
 import { Exponential } from "./nodes/math/exponential/Exponential";
 import { HyperbolicCosine } from "./nodes/math/hyperbolic/HyperbolicCosine";
 import { HyperbolicTangent } from "./nodes/math/hyperbolic/HyperbolicTangent";
-import { IInteractivityVariable, IInteractivityEvent, IInteractivityValue, IInteractivityFlow, IInteractivityGraph, IInteractivityNode, IInteractivityValueType, IInteractivityDeclaration } from "./types/InteractivityGraph";
+import { IInteractivityVariable, IInteractivityEvent, IInteractivityValue, IInteractivityFlow, IInteractivityNode, IInteractivityValueType, IInteractivityDeclaration } from "./types/InteractivityGraph";
 import { VariableInterpolate } from "./nodes/variable/VariableInterpolate";
 import { NoOpNode } from "./nodes/experimental/NoOp";
 import { MatDecompose } from "./nodes/math/matrix/matDecompose";
@@ -117,6 +117,7 @@ import { MathSwitch } from "./nodes/math/special/MathSwitch";
 import { Inverse } from "./nodes/math/matrix/Inverse";
 import { DebugLog } from "./nodes/experimental/Debug";
 import { QuatAngleBetween } from "./nodes/math/quaternion/QuatAngleBetween";
+import { cubicBezier, linearFloat, slerpFloat4 } from "./easingUtils";
 
 
 export class BasicBehaveEngine implements IBehaveEngine {
@@ -364,10 +365,6 @@ export class BasicBehaveEngine implements IBehaveEngine {
         }
     }
 
-    public animateProperty = (path: string, easingParameters: any, callback: () => void) => {
-        //pass
-    }
-
     public animateCubicBezier = (
         path: string,
         p1: number[],
@@ -378,7 +375,40 @@ export class BasicBehaveEngine implements IBehaveEngine {
         valueType: string,
         callback: () => void
     ) => {
-        //pass
+        this.clearPointerInterpolation(path);
+        const startTime = performance.now();
+
+        const action = async () => {
+            const elapsedDuration = (performance.now() - startTime) / 1000;
+            const t = Math.min(elapsedDuration / duration, 1);
+            const p = cubicBezier(t, {x: 0, y:0}, {x: p1[0], y:p1[1]}, {x: p2[0], y:p2[1]}, {x: 1, y:1});
+            if (valueType === "float3") {
+                const value = [linearFloat(p.y, initialValue[0], targetValue[0]), linearFloat(p.y, initialValue[1], targetValue[1]), linearFloat(p.y, initialValue[2], targetValue[2])];
+                this.setPathValue(path, value);
+            } else if (valueType === "float4") {
+                if (this.isSlerpPath(path)) {
+                    const value = slerpFloat4(p.y, initialValue, targetValue);
+                    this.setPathValue(path, value);
+                } else {
+                    const value = [linearFloat(p.y, initialValue[0], targetValue[0]), linearFloat(p.y, initialValue[1], targetValue[1]), linearFloat(p.y, initialValue[2], targetValue[2]), linearFloat(p.y, initialValue[3], targetValue[3])];
+                    this.setPathValue(path, value);
+                }
+            } else if (valueType === "float") {
+                const value = [linearFloat(p.y, initialValue[0], targetValue[0])];
+                this.setPathValue(path, value);
+            } else if (valueType == "float2") {
+                const value = [linearFloat(p.y, initialValue[0], targetValue[0]), linearFloat(p.y, initialValue[1], targetValue[1])];
+                this.setPathValue(path, value);
+            }
+
+            if (elapsedDuration >= duration) {
+                this.setPathValue(path, targetValue);
+                this.clearPointerInterpolation(path);
+                callback();
+            }
+        };
+
+        this.setPointerInterpolationCallback(path, {action: action} );
     }
 
     private registerKnownBehaviorNodes = () => {
